# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IPN-DEV — Sistema de Inventario Daluzed Pastelería",
        "description": "API REST para gestión de inventario, producción y logística de Daluzed Pastelería. Proyecto Nuclear 3 · Corporación Universitaria Alexander von Humboldt (CUE).",
        "contact": {
            "name": "Equipo IPN-DEV",
            "email": "admin@inventario.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/produccion/ordenes": {
            "get": {
                "tags": [
                    "Producción"
                ],
                "summary": "Listar órdenes de producción",
                "operationId": "a9ede6435040db6c8762b2b22bc78a8b",
                "responses": {
                    "200": {
                        "description": "Listado de órdenes"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permiso (requiere produccion.leer)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Producción"
                ],
                "summary": "Crear orden de producción",
                "description": "Etapa 1: planifica la producción de un producto terminado (RFPROD01).",
                "operationId": "78b081c956fd40dc70fab779647d9ad6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "producto_terminado_id",
                                    "cantidad_planificada",
                                    "fecha_planificada"
                                ],
                                "properties": {
                                    "producto_terminado_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "cantidad_planificada": {
                                        "type": "number",
                                        "example": 100
                                    },
                                    "fecha_planificada": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Orden creada en estado pendiente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permiso (requiere produccion.escribir)"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/produccion/ordenes/{id}": {
            "get": {
                "tags": [
                    "Producción"
                ],
                "summary": "Ver orden de producción",
                "operationId": "201e631d56910b4e0b4351dab56f54fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de la orden con consumos"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permiso"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/produccion/ordenes/{id}/ejecutar": {
            "post": {
                "tags": [
                    "Producción"
                ],
                "summary": "Ejecutar orden de producción",
                "description": "Etapa 2: descuenta MP de Bodega Principal usando selección FEFO (RFPROD02). Rechaza si falta stock.",
                "operationId": "49082467cafb02fa7eeee741625c3473",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cantidad_producida"
                                ],
                                "properties": {
                                    "cantidad_producida": {
                                        "type": "number",
                                        "example": 95
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producción ejecutada, MP descontada, lote PT creado en Bodega Producción"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permiso"
                    },
                    "422": {
                        "description": "Stock insuficiente — indica qué MP falta y cuánta (RFPROD05)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/produccion/ordenes/{id}/traslado-pt": {
            "post": {
                "tags": [
                    "Producción"
                ],
                "summary": "Trasladar PT a Bodega Ventas",
                "description": "Etapa 3: mueve el producto terminado de Bodega Producción a Bodega Ventas para despacho (RFPROD03).",
                "operationId": "a7cce888f1e64931130851a40585de0c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PT trasladado a Bodega Ventas"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permiso"
                    },
                    "422": {
                        "description": "Orden no ejecutada o ya trasladada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/produccion/ordenes/{id}/anular": {
            "patch": {
                "tags": [
                    "Producción"
                ],
                "summary": "Anular orden de producción",
                "description": "Crea movimientos compensatorios inmutables. No borra registros (HU-027).",
                "operationId": "3fcc9bcdea8826d5272d87d9bcd74d5e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orden anulada con movimiento compensatorio"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permiso"
                    },
                    "422": {
                        "description": "No se puede anular en el estado actual"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token Bearer generado por Laravel Sanctum al hacer login. Incluirlo en el header: Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}